Restructure to allow lib to be initialized with socket obj
To further remove the code needed to boot ratchetjs, the channel logic
was moved over. However the socket connection is setup in app, so an
interface was needed to "init" ratchet with a socket to tie things up.
This is is a mess. I recognize that and feel shame :D