Replace for loop with iteration
I really hate to fix this without a test, but... :sweat_smile:
There was a frustratingly subtle bug with the for loop. The anonymous
function defined in the loop would close on the variables as references.
The `wrapper` was a particular problem. Basically whatever the last
wrapper was would be the one used to render _all_ components on the
page. This is fixed by using iteration, because each element is yielded
to a function which has its own scope. OOF